feat(panels): reorder multiple fill layers
