+++ crates/op-editor-core/src/mutators.rs
@@ -13,7 +13,6 @@
 //! mutators stay page-model-agnostic.
 
 use crate::geometry::{union_aggregate_bounds, DocRect};
-use crate::history::EditorSnapshot;
 use crate::node_id::NodeId;
 use crate::pen_node_ext::PenNodeExt;
 use crate::selection::SelectionState;
@@ -21,9 +20,6 @@ use crate::state::EditorState;
 use crate::walkers::{self, find_node, find_node_mut, reorder_in_children, ReorderDirection};
 use jian_ops_schema::node::PenNode;
 use std::collections::HashSet;
-
-/// Largest number of undo entries kept (matches shell-core's cap).
-const HISTORY_CAP: usize = 100;
 
 impl EditorState {
     // --- Active-page node access -------------------------------------
@@ -255,116 +251,6 @@ impl EditorState {
         }
         self.selection.anchor = ids.last().cloned().unwrap_or(NodeId::NONE);
         self.selection.set = ids;
-        true
-    }
-
-    // --- History -----------------------------------------------------
-
-    /// Snapshot the editor's undoable state without pushing it.
-    ///
-    /// The snapshot covers `doc` / `selection` / `active_page_index`
-    /// only. View-only UI state — notably `editor_ui.collapsed_layers`
-    /// — is intentionally NOT captured: layer-collapse is a view-only
-    /// toggle, deliberately excluded from the undo snapshot and from
-    /// file persistence. Expanding / collapsing a layer is not an
-    /// undoable edit.
-    pub fn snapshot_for_history(&self) -> EditorSnapshot {
-        // Default anchor: the adjacent history state. After an undo the
-        // redo stack is non-empty and its back is the state we came
-        // FROM (the closest neighbour of the current live state) — so a
-        // divergent edit captured here shares subtrees with it, and the
-        // anchor is read BEFORE `history_push_past` clears the redo
-        // stack. In normal forward editing the redo stack is empty, so
-        // the anchor falls through to the previous undo entry.
-        let anchor = self
-            .history
-            .future
-            .back()
-            .or_else(|| self.history.past.back());
-        self.snapshot_for_history_with_anchor(anchor)
-    }
-
-    /// [`snapshot_for_history`](Self::snapshot_for_history) with an
-    /// explicit anchor. `anchor` is the adjacent history snapshot whose
-    /// unchanged top-level subtrees the new snapshot should share by
-    /// `Arc`; `None` produces an all-fresh snapshot. Undo / redo pass
-    /// the popped destination here so the parked entry shares with
-    /// where the editor is going.
-    pub fn snapshot_for_history_with_anchor(
-        &self,
-        anchor: Option<&EditorSnapshot>,
-    ) -> EditorSnapshot {
-        EditorSnapshot {
-            doc: crate::history_snapshot::SharedDoc::capture(&self.doc, anchor.map(|s| &s.doc)),
-            selection: self.selection.clone(),
-            active_page_index: self.ui.active_page_index,
-            components: crate::history_snapshot::SharedComponents::capture(
-                &self.components,
-                anchor.map(|s| &s.components),
-            ),
-            app_state_owner: self.app_state_owner.clone(),
-            revision: self.revision,
-        }
-    }
-
-    /// Push a snapshot onto the undo stack + clear redo. Cap = 100.
-    pub fn history_push_past(&mut self, snap: EditorSnapshot) {
-        self.history_push_count = self.history_push_count.saturating_add(1);
-        self.history.past.push_back(snap);
-        if self.history.past.len() > HISTORY_CAP {
-            self.history.past.pop_front();
-        }
-        self.history.future.clear();
-        self.mark_document_changed();
-    }
-
-    /// Push the current state onto the undo stack. Call BEFORE a
-    /// transactional change so undo reverts to here.
-    pub fn commit_history(&mut self) {
-        let snap = self.snapshot_for_history();
-        self.history_push_past(snap);
-    }
-
-    /// Restore the editor state from a snapshot. Materializes the
-    /// shared document + components back into owned values. `pub(crate)`
-    /// so batch rollback ([`crate::command_batch`]) routes through the
-    /// same materializing path instead of moving shared state into live
-    /// fields.
-    pub(crate) fn restore(&mut self, snap: EditorSnapshot) {
-        self.doc = snap.doc.materialize();
-        self.selection = snap.selection;
-        self.ui.active_page_index = snap.active_page_index;
-        self.components = snap.components.materialize();
-        self.app_state_owner = snap.app_state_owner;
-        self.revision = snap.revision;
-        self.sync_dirty_flag();
-    }
-
-    /// Undo the last change. False when the undo stack is empty.
-    pub fn undo(&mut self) -> bool {
-        let Some(prev) = self.history.past.pop_back() else {
-            return false;
-        };
-        // Capture the current (live) state as the redo entry, anchored
-        // on the POPPED destination — before it is materialized — so the
-        // parked snapshot shares every unchanged subtree with where the
-        // editor is about to land.
-        let cur = self.snapshot_for_history_with_anchor(Some(&prev));
-        self.history.future.push_back(cur);
-        self.restore(prev);
-        true
-    }
-
-    /// Redo the last undone change. False when the redo stack is empty.
-    pub fn redo(&mut self) -> bool {
-        let Some(next) = self.history.future.pop_back() else {
-            return false;
-        };
-        // Symmetric to undo: anchor the parked undo entry on the popped
-        // redo destination.
-        let cur = self.snapshot_for_history_with_anchor(Some(&next));
-        self.history.past.push_back(cur);
-        self.restore(next);
         true
     }
